Reorganize spacecraft by organization
